refactor: moved json ops to a separate class
