+++ benchmark/Insight.Database.Benchmarks.csproj
@@ -3,6 +3,8 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Insight.Database.Benchmark</RootNamespace>
+    <AssemblyName>Insight.Database.Benchmarks</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
